Annotations can now be populated with any query result, where entries contain a @timestamp field
